backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
openai-whisper==20230314
transformers==4.35.0
torch==2.0.1
accelerate==0.23.0
sentencepiece==0.1.99
numpy==1.23.5
scipy==1.11.3
soundfile==0.12.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.8
alembic==1.12.0
bert-score==0.3.13
rouge-score==0.1.2
python-dotenv==1.0.0
